# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Directioner API specification
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export type ProfileTier = typeof ProfileTier[keyof typeof ProfileTier];


export const ProfileTier = {
  free: 'free',
  basic: 'basic',
  pro: 'pro',
  max: 'max',
} as const;

export interface Profile {
  id: string;
  username: string;
  tier: ProfileTier;
  creditsUsed: number;
  creditsLimit: number;
  createdAt: string;
  updatedAt: string;
}

export type UpdateProfileTier = typeof UpdateProfileTier[keyof typeof UpdateProfileTier];


export const UpdateProfileTier = {
  free: 'free',
  basic: 'basic',
  pro: 'pro',
  max: 'max',
} as const;

export interface UpdateProfile {
  username?: string;
  tier?: UpdateProfileTier;
  creditsUsed?: number;
  creditsLimit?: number;
}

export type ServerStatus = typeof ServerStatus[keyof typeof ServerStatus];


export const ServerStatus = {
  online: 'online',
  offline: 'offline',
  maintenance: 'maintenance',
} as const;

export interface Server {
  id: string;
  userId: string;
  serverName: string;
  discordServerId: string;
  memberCount: number;
  channelCount: number;
  status: ServerStatus;
  tier: string;
  aiMode: string;
  createdAt: string;
}

export type CreateServerStatus = typeof CreateServerStatus[keyof typeof CreateServerStatus];


export const CreateServerStatus = {
  online: 'online',
  offline: 'offline',
  maintenance: 'maintenance',
} as const;

export interface CreateServer {
  serverName: string;
  discordServerId: string;
  memberCount?: number;
  channelCount?: number;
  status?: CreateServerStatus;
  tier?: string;
  aiMode?: string;
}

export type UpdateServerStatus = typeof UpdateServerStatus[keyof typeof UpdateServerStatus];


export const UpdateServerStatus = {
  online: 'online',
  offline: 'offline',
  maintenance: 'maintenance',
} as const;

export interface UpdateServer {
  serverName?: string;
  discordServerId?: string;
  memberCount?: number;
  channelCount?: number;
  status?: UpdateServerStatus;
  tier?: string;
  aiMode?: string;
}

export interface AnalyticsEntry {
  id: string;
  userId: string;
  serverId?: string | null;
  date: string;
  textMessages: number;
  voiceMinutes: number;
  credits: number;
  createdAt: string;
}

export interface CreateAnalytics {
  serverId?: string;
  date: string;
  textMessages?: number;
  voiceMinutes?: number;
  credits?: number;
}

export type MemoryNodeScope = typeof MemoryNodeScope[keyof typeof MemoryNodeScope];


export const MemoryNodeScope = {
  user: 'user',
  server: 'server',
  global: 'global',
} as const;

export interface MemoryNode {
  id: string;
  userId: string;
  serverId?: string | null;
  content: string;
  scope: MemoryNodeScope;
  targetUser?: string | null;
  createdAt: string;
}

export type CreateMemoryNodeScope = typeof CreateMemoryNodeScope[keyof typeof CreateMemoryNodeScope];


export const CreateMemoryNodeScope = {
  user: 'user',
  server: 'server',
  global: 'global',
} as const;

export interface CreateMemoryNode {
  serverId?: string;
  content: string;
  scope?: CreateMemoryNodeScope;
  targetUser?: string;
}

export interface ActivityEntry {
  id: string;
  userId: string;
  serverId?: string | null;
  eventType: string;
  channel?: string | null;
  actor?: string | null;
  details: string;
  createdAt: string;
}

export interface CreateActivityEntry {
  serverId?: string;
  eventType: string;
  channel?: string;
  actor?: string;
  details?: string;
}

export type BillingEntryStatus = typeof BillingEntryStatus[keyof typeof BillingEntryStatus];


export const BillingEntryStatus = {
  paid: 'paid',
  pending: 'pending',
  failed: 'failed',
  refunded: 'refunded',
} as const;

export interface BillingEntry {
  id: string;
  userId: string;
  amount: string;
  status: BillingEntryStatus;
  description?: string | null;
  invoiceUrl?: string | null;
  createdAt: string;
}

export type CreateBillingEntryStatus = typeof CreateBillingEntryStatus[keyof typeof CreateBillingEntryStatus];


export const CreateBillingEntryStatus = {
  paid: 'paid',
  pending: 'pending',
  failed: 'failed',
  refunded: 'refunded',
} as const;

export interface CreateBillingEntry {
  amount: string;
  status?: CreateBillingEntryStatus;
  description?: string;
  invoiceUrl?: string;
}

export type ListAnalyticsParams = {
serverId?: string;
from?: string;
to?: string;
};

export type ListMemoryNodesParams = {
serverId?: string;
};

export type ListActivityFeedParams = {
serverId?: string;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};