# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Directioner API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analytics:
    get:
      operationId: listAnalytics
      tags: [analytics]
      summary: List analytics entries
      security:
        - bearerAuth: []
      parameters:
        - name: serverId
          in: query
          schema:
            type: string
            format: uuid
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Analytics entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnalyticsEntry"
    post:
      operationId: createAnalytics
      tags: [analytics]
      summary: Record analytics entry
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnalytics"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsEntry"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AnalyticsEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        serverId:
          type: string
          format: uuid
          nullable: true
        date:
          type: string
          format: date
        textMessages:
          type: integer
        voiceMinutes:
          type: integer
        credits:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - date
        - textMessages
        - voiceMinutes
        - credits
        - createdAt
    CreateAnalytics:
      type: object
      properties:
        serverId:
          type: string
          format: uuid
        date:
          type: string
          format: date
        textMessages:
          type: integer
        voiceMinutes:
          type: integer
        credits:
          type: integer
      required:
        - date
...